refactor(pages): refactor home, event, volunteer pages
