--- src/App.tsx
@@ -1,10 +1,0 @@
-import EventPage from "./pages/Events/EventPage.tsx";
-function App() {
-  return (
-    <>
-      <EventPage />
-    </>
-  );
-}
-
-export default App;
