+++ src/components/layout/Navbar.tsx
@@ -15,7 +15,7 @@ const BloodDonationNavbar = () => {
   const navItems: NavItem[] = [
     { id: "su-kien", label: "Sự kiện hiến máu", href: "/events" },
     { id: "trang-chu", label: "Trang chủ", href: "/" },
-    { id: "ve-chung-toi", label: "Về chúng tôi", href: "/about" },
+    { id: "ve-chung-toi", label: "Thông tin máu", href: "/about" },
     { id: "chia-se", label: "Chia sẻ", href: "/share" },
   ];
