--- src/pages/DonationEvent/VolunteerPage.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function VolunteerPage() {
-  return (
-    <div>
-      <h1>Volunteer page</h1>
-    </div>
-  )
-}
-
-export default VolunteerPage
