--- src/pages/Events/EventPage.tsx
@@ -1,15 +1,0 @@
-import Events from "@/pages/Events/Events";
-import Footer from "@/components/layout/Footer";
-import BloodDonationNavbar from "@/components/layout/Navbar";
-
-function EventPage() {
-  return (
-    <>
-      <BloodDonationNavbar />
-      <Events />
-      <Footer />
-    </>
-  );
-}
-
-export default EventPage;
